Move the agsmaster relevant code into its own file.
